move ffcrawler class to ligpargen.py
