+++ mdgo/forcefield/__init__.py
@@ -16,7 +16,6 @@ __date__ = "Dec 19, 2023"
 
 from .aqueous import IonLJData, Aqueous
 from .charge import ChargeWriter
-from .crawler import FFcrawler
 from .maestro import MaestroRunner
 from .pubchem import PubChemRunner
-from .ligpargen import LigpargenRunner
+from .ligpargen import LigpargenRunner, FFcrawler
